# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: DocFlow API
  version: 1.0.0
  description: |
    DocFlow 文档管理服务 API。

    ## 认证
    - 业务端点要求 `Authorization: Bearer <access_token>`（HS256 JWT，subject 为用户 UUID）。
    - 登录/刷新通过 HttpOnly Cookie `refresh_token`（Path=/api/v1/auth/refresh，SameSite=Lax）轮换会话。

    ## 错误与限流
    - 错误响应统一为 `{error, code?}`；`code` 为机器可读错误码（如 NAME_CONFLICT、QUARANTINED）。
    - 认证接口按 user_id+IP 每分钟限流；登录按 IP+邮箱前缀哈希更严格限流；
      公开分享接口按 IP 独立限流；超限返回 429 并带 Retry-After。
    - 文件元数据响应带 ETag（基于 updated_at）。
servers:
  - url: /api/v1
    description: API 基础路径（契约测试据此归一化路由）
security:
  - bearerAuth: []
...
paths:
...
  /public/shares/{token}/preview:
    get:
      tags: [Public Shares]
      operationId: publicSharePreview
      summary: 公开分享预览（无认证）
      description: >-
        view 与 download 权限均可预览；不消耗分享 download_count；
        MIME 白名单与安全头同个人预览。网页包联动同个人预览：zip 且存在
        ready 解包结果时返回 200 JSON（WebpkgPreview）。预览成功时写入
        访问事件（action=preview）。密码保护分享须先经 verify 取得会话
        cookie（未携带时 401 code=PASSWORD_REQUIRED）。
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
        - name: Range
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: 内联内容；zip 网页包为 WebpkgPreview JSON
          headers:
            Content-Disposition: { schema: { type: string } }
            X-Content-Type-Options: { schema: { type: string, const: nosniff } }
            Accept-Ranges: { schema: { type: string, const: bytes } }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
            application/json:
              schema: { $ref: '#/components/schemas/WebpkgPreview' }
        '206':
          description: 区间内容
          headers:
            Content-Range: { schema: { type: string } }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '401':
          description: 分享受密码保护且未携带有效会话 cookie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 对象不可用或分享不允许访问
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { $ref: '#/components/responses/Gone' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '416':
          description: Range 无法满足
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /public/shares/{token}/office:
    get:
      tags: [Public Shares]
      operationId: publicShareOfficeConfig
      summary: 公开分享 OnlyOffice 查看会话（无认证）
      description: >-
        Office 文档（doc/docx/xls/xlsx/ppt/pptx/odt/ods/odp）访客在线查看：
        解析分享 token（有效期/撤销/密码会话同 preview 口径），返回可直接
        传给 DocsAPI.DocEditor 的 JWT 签名只读配置（恒 view 模式；分享
        permission=download 时编辑器内开放下载/打印）与浏览器可达的
        DocumentServer 地址 server_url。view 与 download 权限均可预览、
        不消耗 download_count；预览成功计入 view_count 与访问事件
        （action=preview）。ONLYOFFICE 集成未启用时 404（前端回退
        「不支持在线预览」）。
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
        - name: lang
          in: query
          required: false
          description: 编辑器界面语言（BCP 47，如 zh-CN；非法值回退 zh）
          schema: { type: string }
      responses:
        '200':
          description: 只读查看会话配置
          content:
            application/json:
              schema:
                type: object
                required: [server_url, config]
                properties:
                  server_url:
                    type: string
                    description: DocumentServer 浏览器可达地址（加载 api.js）
                  config:
                    type: object
                    description: JWT 签名的 DocEditor 配置（document/editorConfig/token）
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: 分享受密码保护且未携带有效会话 cookie
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: 对象不可用或分享不允许访问
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { $ref: '#/components/responses/Gone' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /onlyoffice/session:
    post:
      tags: [onlyoffice]
      operationId: createOnlyOfficeSession
      summary: 生成编辑器配置（DocsAPI.DocEditor 可直接使用）
      description: >-
        校验读权限（authorizeFileAccess 同规则：个人文件 owner、团队文件任意
        在册成员）且当前版本 blob 为 available。editorConfig.mode 按写权限判定：
        可写（个人 owner / 团队 CanWrite，与「覆盖为新版本」目标校验同语义）时
        为 edit 且 document.permissions.edit=true；只读用户为 view 且
        permissions.edit=false（服务端未接线写授权器时保守降级 view）。
        document.key 绑定 `file_id:version_id`（编辑会话与版本绑定）；
        document.url 为 5 分钟有效期的签名下载 URL
        （ONLYOFFICE_DOWNLOAD_URL_BASE 前缀）；token 为对 document+
        editorConfig 整体签名的 JWT（HS256，ONLYOFFICE_JWT_SECRET，TTL 5 分钟，
        aud=onlyoffice-config，回调端点拒绝该用途 token）。集成未启用
        （ONLYOFFICE_ENABLED=false）时路由不注册（404）。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [file_id]
              properties:
                file_id: { type: string, format: uuid }
      responses:
        '200':
          description: >-
            DocsAPI.DocEditor 配置（documentType/document/editorConfig/token；
            editorConfig 含 mode（edit|view，按写权限判定）、user.id/name、
            lang=zh 与 callbackUrl；document.permissions.edit 与 mode 一致）
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403':
          description: 无权限或当前版本不可用（非 available 状态）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: 文件无当前版本
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /onlyoffice/download/{fileId}:
    get:
      tags: [onlyoffice]
      operationId: onlyofficeDownload
      summary: DocumentServer 回源下载（签名 token，无 Bearer）
      description: >-
        供 DocumentServer 拉取文档内容（无用户会话）。token 为 HS256 JWT：
        sub=file_id、v=版本 ID、aud=onlyoffice-download、exp 5 分钟，且须与
        路径 fileId、查询 v 一致；校验通过后流式返回该版本内容（单区间 Range 与
        Content-Disposition: attachment 同个人下载，不计入 download_count）。
        被隔离/不可用 blob 一律拒绝（403）。独立按 IP 轻限流
        （ONLYOFFICE_RATE_LIMIT_PER_MIN）。
      security: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: token
          in: query
          required: true
          schema: { type: string }
          description: 签名下载 token（session 接口的 document.url 内嵌）
        - name: v
          in: query
          required: true
          schema: { type: string, format: uuid }
          description: 版本记录 ID（与 token 的 v 声明一致）
        - name: Range
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: 全量内容
          headers:
            Content-Disposition: { schema: { type: string } }
            Accept-Ranges: { schema: { type: string, const: bytes } }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '206':
          description: 区间内容
          headers:
            Content-Range: { schema: { type: string } }
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401':
          description: token 缺失、签名不符、过期、用途（aud）不符或参数绑定失败
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: blob 非 available 状态（隔离/不可用）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '416':
          description: Range 无法满足
          headers:
            Content-Range: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /onlyoffice/download/{fileId}/{filename}:
    get:
      tags: [onlyoffice]
      operationId: onlyofficeDownloadNamed
      summary: DocumentServer 回源下载（带文件名兼容路由）
      security: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: filename
          in: path
          required: true
          schema: { type: string }
        - name: v
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: token
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: 文件内容 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /onlyoffice/callback:
    post:
      tags: [onlyoffice]
      operationId: onlyofficeCallback
      summary: 保存回调（DocumentServer → DocFlow，无 Bearer）
      description: >-
        DocumentServer command service 回调。JWT 校验取 body.token 或
        Authorization: Bearer 头（HS256，ONLYOFFICE_JWT_SECRET）。token 用途
        隔离：aud 为 onlyoffice-config 或 onlyoffice-download 的 token 一律
        拒绝（编辑配置/下载 token 不得混用为回调 token；DS 自签回调 token
        通常无 aud，允许通过）。status/key/url 以验签后的 claims 为唯一信任
        源：claims 缺失必需声明（status、key，保存类状态另需 url）即拒绝，
        不回退未签名 body。status=2/6：先做写鉴权（claims users[0] 为 actor，
        按「覆盖为新版本」目标校验同语义判定写权限；未接线授权器时 fail
        closed 拒绝），通过后以幂等键 (file_id, document.key, url) 抢占
        （重复回调不重复建版本），仅接受与 ONLYOFFICE_SERVER_URL 同源的下载
        URL（scheme/host/port 全等，防 SSRF），下载后作为新版本保存（临时
        key 转正、SHA-256 校验、复用内容去重）。
        status=4 仅清理（写 onlyoffice.cleanup 审计）；其他状态空处理。
        始终返回 HTTP 200：成功 {"error":0}，失败 {"error":1}（协议约定，
        非 200 会触发 DocumentServer 重试）。独立按 IP 轻限流。
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                key: { type: string, description: document.key（file_id:version_id） }
                status: { type: integer, description: '1=编辑中 2=保存 3=错误 4=关闭 6=强制保存 7=强制保存错误' }
                url: { type: string, description: 新版本下载 URL（仅 status=2/6 使用） }
                users: { type: array, items: { type: string } }
                token: { type: string, description: 回调体整体签名的 JWT }
      responses:
        '200':
          description: '处理结果（协议体）：成功 error=0，失败 error=1'
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: integer, enum: [0, 1] }
        '429': { $ref: '#/components/responses/RateLimited' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Authorization: Bearer <token>，两类凭证等价可用：
        ① HS256 JWT（subject 为用户 UUID，过期时间以服务配置为准）；
        ② PAT 个人访问令牌：dfpat_ + 43 字符 URL-safe base64（经
        POST /tokens 创建，撤销/过期即失效；无 refresh 语义）。
  responses:
    BadRequest:
      description: 请求参数或请求体非法
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: 缺失或无效的访问令牌
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: 资源不存在或不属于当前用户（不泄露存在性）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Gone:
      description: 资源已失效（分享过期/撤销/达下载上限、上传会话过期或终态）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    UnsupportedMediaType:
      description: Content-Type 不被接受，或预览 MIME 不在白名单（preview not supported）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: 触发限流；响应头 Retry-After 为当前窗口剩余秒数
      headers:
        Retry-After:
          description: 建议等待秒数
          schema: { type: integer }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: 人类可读错误信息
        code:
          type: string
          description: >-
            机器可读错误码；当前实现使用 NAME_CONFLICT / NOT_DELETED / PARENT_DELETED /
            QUARANTINED / FAILED / TOTP_REQUIRED（POST /auth/login 对已启用两步
            验证账号的 401，error=totp_required）/ ACCOUNT_LOCKED（登录与
            refresh 遇锁定期账号的 423）/ QUOTA_EXCEEDED（上传建会话配额
            超限的 403）/ CSRF_REJECTED（refresh/logout 同源校验失败的 403）
        status:
          type: string
          description: 上传相关错误附带的会话状态（仅 uploads complete 端点）
...
    WebpkgPreview:
      type: object
      description: >-
        网页包预览指引（/files/{id}/preview 与 /public/shares/{token}/preview
        对 zip+ready 包的 200 JSON 响应）；前端以 sandbox="allow-scripts" 的
        iframe 加载 url（无 allow-same-origin，内容端点带严格 CSP/nosniff）。
      required: [kind, url]
      properties:
        kind: { type: string, const: webpkg }
        url:
          type: string
          description: '内容入口 /content/<public_id>/index.html（public_id 为 43 字符随机串）'
...